Created heron/spouts as the location to house all well known spouts Moved TextFile/FixedLines spouts/streamlets to seed it
