--- heron/dsl/src/python/inputs/inmemory/__init__.py
@@ -1,3 +1,0 @@
-'''DSL module for simple inmemory static line inputs'''
-
-from .staticlines import StaticLinesStreamlet
