--- heron/dsl/src/python/inputs/textfile/__init__.py
@@ -1,3 +1,0 @@
-'''DSL module for textfile inputs'''
-
-from .textfile import TextFileStreamlet
