+++ heron/spouts/src/python/__init__.py
@@ -1,0 +1,11 @@
+"""
+The top-level library for Heron's Python Spouts, which define some basic
+Heron Spouts/Streamlets to be used while constructing Heron Topologies either
+using the basic Low Level API and the higher level DSL API.
+"""
+
+# Load all spout modules
+from .fixedlines.fixedlinesspout import FixedLinesSpout
+from .fixedlines.fixedlinesstreamlet import FixedLinesStreamlet
+from .textfiles.textfilespout import TextFileSpout
+from .textfiles.textfilestreamlet import TextFileStreamlet
