+++ heron/spouts/src/python/textfiles/textfilespout.py
@@ -11,15 +11,10 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-'''textfile.py: module for defining text file input'''
-import glob
-
+'''textfilespout.py: module that defines a Heron Spout that reads data
+   from a list of files and emits one tuple per line'''
 from heron.api.src.python import Spout
 
-from ...streamlet import Streamlet, OperationType
-
-# pylint: disable=access-member-before-definition
-# pylint: disable=attribute-defined-outside-init
 class TextFileSpout(Spout):
   """TextFileSpout: reads from a list of files"""
 
@@ -90,30 +85,3 @@ class TextFileSpout(Spout):
   def fail(self, tup_id):
     self.fail_count += 1
     self.logger.debug("Failed tuple %s" % str(tup_id))
-
-class TextFileStreamlet(Streamlet):
-  """A TextFileStreamlet is a list of text input files
-  """
-  def __init__(self, filepattern, stage_name=None, parallelism=None):
-    self._files = glob.glob(filepattern)
-    super(TextFileStreamlet, self).__init__(operation=OperationType.Input,
-                                            stage_name=stage_name,
-                                            parallelism=parallelism)
-
-  @staticmethod
-  def textFile(filepattern, stage_name=None, parallelism=None):
-    return TextFileStreamlet(filepattern, stage_name=stage_name, parallelism=parallelism)
-
-  def _build(self, bldr, stage_names):
-    self._parallelism = len(self._files)
-    if self._parallelism < 1:
-      raise RuntimeError("No matching files")
-    if self._stage_name is None:
-      index = 1
-      self._stage_name = "input"
-      while self._stage_name in stage_names:
-        index = index + 1
-        self._stage_name = "input" + str(index)
-      bldr.add_spout(self._stage_name, TextFileSpout, par=self._parallelism,
-                     config={TextFileSpout.FILES : self._files})
-    return bldr
